Move route registration to separate file
